tdf#161001 sw floattable: require in-header/footer for non-follow-text-flow
Open the bugdoc, the in-body floating table has an inner floating table
which is rendered above the outer table, not inside it.

This went wrong in commit f74a6ef94ac957e4c146fc9923d30ce6bd31b5ce
(tdf#159453 sw floattable: fix unexpected overlap of in-header fly and
body text, 2024-02-13), where the use-case was an in-header floating
table interacting with body text.

The more general case is in-body floating tables interacting with
in-body text, so fix the problem by restricting the custom
IsFollowingTextFlow=false property for in-header floating tables.

Change-Id: I902f232e3c840ff0279d8ed6c2329ba48a03b4e1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180534
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins