tdf#164785: restore saving/reapplying of CF keys in cells
Commit c492de66a077f3a2a960209b0b8b278b3901f361 (tdf#160149: save and
restore the whole set of tab's conditional formats, 2024-03-17) had
replaced copying of the affected cell ranges (including respective CF)
with copying of the whole CF list. That allowed to replace the list as
a whole, and be sure that it is in the same state as before the undone
operation (not merged in any way because of re-application of CF).

But the previous code made sure to set the CF keys in affected cells.
New code didn't do that, and so, the changes to the cell attributes with
CF keys were not cleared on undo. Applying new CF to another cell range
created the same key that was already referenced in those old cells, so
they behaved as if they belong to the new CF range.

This change keeps the "save/restore whole CF list" idea, so tdf#160149
is kept fixed. But additionally, it restores saving/reapplying of cell
attributes (but not other data) in the areas affected by CF.

Change-Id: I4ff844ca1345322e231df605fb019b9613a6b123
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180555
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>