tdf#130857 qt weld: Support custom MenuButton popover
Like GtkMenuButton [1], weld::MenuButton doesn't
only support showing a menu, but also a custom
popover/widget.

This widget can be set via weld::MenuButton::set_popover.
Implement that for QtInstanceMenuButton and adjust the
logic to use either that popover (if set) or a menu
set for the underlying QToolButton.

Manually trigger either showing the popover or the menu
in a slot connected to the QToolButton::clicked signal.

One dialog using menu buttons with custom popovers
is the one that can be triggered in Writer via "Format" -> "Theme",
then clicking the "Add" button in that dialog.
(That dialog isn't supported by QtInstanceBuilder yet,
however, just seen that while experimenting.)

Change-Id: I9a5924a5ed8c5b8b39050849fe609188559c479c
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180561
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins