tdf#160196 add support for PDF/UA-2 and NS for struct. elem.
This adds support for PDF/UA-2 which is used when using PDF 2.0
(and PDF/UA-1 when using PDF 1.x).

PDF/UA-2 requires that the PDF 2.0 introduced namespaces are set
for the structure elements, so this also adds support for the
namespaces with always setting the PDF 2.0 namespace for all the
structured elements (we don't use others anyway).

Change-Id: Ifbd8f8126d9bcc0d4dcd06a5f56bacdb0804adbd
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180565
Reviewed-by: Tomaž Vajngerl <quikee@gmail.com>
Tested-by: Jenkins
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180876
Tested-by: Jenkins CollaboraOffice <jenkinscollaboraoffice@gmail.com>
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>