cool#10630 lok doc sign: reduce the default size of the signature line
This is obviously somewhat subjective, but the default shape size in
officecfg/registry/schema/org/openoffice/Office/Impress.xcs,
DefaultObjectSize is 8x5cm, and I find myself always reduring the
inserted signature line when I test it on a few sample PDF files where
others left some space for a signature.

Improve this by going with the half of the default in case of signature
lines, so people typically need to only move the inserted shape, not
resize it.

The need to only move it is especially helpful as our resize would not
keep aspect ratio by default and only few users will know they need to
hold ctrl-shift while resizing to do proportional resize.

Change-Id: Ieaad56601be8277b39a5403e3c3f2e6ed1fd1c97
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180587
Reviewed-by: Miklos Vajna <vmiklos@collabora.com>
Tested-by: Jenkins