tdf#130857 qt weld: Actually paint custom bg color
Call QWidget::setAutoFillBackground(true) [1] so
that the color set for the widget background is
actually automatically filled with the color that was set.

Stop setting a custom role for the background role.
Instead, set the custom color for the role that
is already set as the background role.

Make sure to run in the main thread.

With this in place, the "warningmsg" label in
the "Security Level" tab
(xmlsecurity/uiconfig/ui/securitylevelpage.ui)
of the macro security dialog now actually shows with
the background color that gets set in
MacroSecurityLevelTP::SetWarningLabel in a WIP
branch where support for that dialog is declared
in QtInstanceBuilder.

[1] https://doc.qt.io/qt-6/qwidget.html#autoFillBackground-prop

Change-Id: Ib09ef25ee8d7b84fb8287f9b7abb25459a708be7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180611
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>