tdf#130857 qt weld: Implement setting font color
Similar to the existing QtInstanceWidget::set_background,
add a helper method QtInstanceWidget::setFontColor
to set the font/foreground color, and call it from the
subclasses that have a set_font_color method.

With this in place, the "warningmsg" label in
the "Security Level" tab
(xmlsecurity/uiconfig/ui/securitylevelpage.ui)
of the macro security dialog now shows with
the font color that gets set in
MacroSecurityLevelTP::SetWarningLabel in a WIP
branch where support for that dialog is declared
in QtInstanceBuilder.

Change-Id: I100d4c708bd6d19e99870cd0c2158ce542326480
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180612
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>