toolkit: Drop ITableInputHandler
Drop the ITableInputHandler interface (abstract class)
that is only implemented by a single subclass,
DefaultInputHandler, and use that one directly.

Adjust the PTableInputHandler to

    typedef std::shared_ptr<DefaultInputHandler> PTableInputHandler

and take over method comments to DefaultInputHandler.

Make methods non-virtual and follow the
loplugin:staticaccess suggestion to make 3 of
the methods static.

This also prepares for further code simplification in
upcoming commits.

Change-Id: Ib4f3c2eb9f132da0484ef86a404251e3cdd84890
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180630
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins