Avoid matching the 'if' with an empty reference
Commit d833992e7e06f7d385fca6ffe35ce66b79360a63  (Avoid checking
exact interface type of Any value,  2024-11-08)  had modified an
'if' condition in ScFunctionAccess::callFunction, so an Any with
a reference  to an object  implementing  the requested interface
could be matched, even if the Any points to another interface of
the object. However, it made an empty XInterface reference match,
too, which is unexpected.

It is not a real problem now:  both no match,  and a match of an
empty reference  set the error;  but in the future,  a new "else"
branch could be added  after this one,  and then it could behave
unexpectedly, so fix it here.

Change-Id: I43375fc34fe01665e7984ea6ad7052f7969c13c0
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180636
Tested-by: Jenkins
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>