tdf#160578: do not modify procedure argument type based on its use
Since the initial import, there is a code in SbiExpression::Term, that
checks if the defined symbol is used as object (e.g., dot notation to
access its members), and then, if its defined type is Variant, then it
is corrected to Object. There is no rationale for this in comments; so
I have no way to know what could break if that core is dropped.

It's obvious that at least for procedure arguments, such correction is
wrong: the argument definition is the procedure's API; and the actual
use of the symbol should not unexpectedly change what is advertised.

This change limits the application of the correction to non-arguments.
If it should be dropped completely is a separate question.

Change-Id: Ia902afa3f744d0d51510ba6903be45f78e2f6429
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180665
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins
(cherry picked from commit f071292a5af10f4718302a1bf9cc9cdc37528225)
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180675
Reviewed-by: Xisco Fauli <xiscofauli@libreoffice.org>