toolkit a11y: Return parent-relative coords if requested
As its documentation described, the (so far) purely virtual
AccessibleGridControlBase::implGetBoundingBox was implemented
by all subclasses to return the bounding box relative to the
TableControl's vcl::Window parent.

However, this method is used to implement methods like
XAccessibleComponent::containsPoint,
XAccessibleComponent::getBounds and
XAccessibleComponent::getLocation, all of which are supposed
to be using the position relative to the *immediate* accessible
parent.

This mismatch could e.g. be seen with the grid control
in attachment 198647 of tdf#164783 and the gtk3 VCL plugin
as follows:

1) allow macros
2) open sample file
3) click "Grid in dialog" to start the dialog containing a grid
4) start Accerciser
5) In Accerciser's treeview of the LO a11y hierarchy, select the
   "Grid control" object with a11y table role.
6) Select its first child, which is a table cell.

What could be seen is that the top-left corner of the highlighted
area for the first cell (step 6) is shifted to the right and down
as compareed to the highlighted area for the table (step 5), even
though the first cell in reality has the exact same origin, (because
the table area is defined by the cells it contains).

The AccessibleGridControlBase::implGetBoundingBoxOnScreen overrides
correctly return the bounding box on screen and duplicated much
of the logic.

Fix the problem and deduplicate code by implementing
AccessibleGridControlBase::implGetBoundingBox directly in
AccessibleGridControlBase: Let it calculate the position
relative to the parent from the absolute position of the parent
and of the object itself.

With this in place, the scenario described above now results in
the area highlighted in steps 5) and 6) to have the same top-left
corner.

Change-Id: If28de14d440706bd6493626586f82859e5570e75
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180720
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins