tdf#164844: handle colorscale and friends specially when deduplicating CF
The conditional formats like colorscale, databar and iconset may take
their whole range to calculate their maximum / minimum, and each cell
format would depend on the CF range. Thus, automatic merge of ranges
of similar CFs when copying/pasting could create unexpected results.
E.g., merging CFs of two distinct color scales - one for a range from
1 to 10, and another from 100 to 500 - would change colors on both of
them, which is usually not what user wants, even when these ranges
are side by side in adjacent columns.

On the other hand, users still may want to have a way to expand the
range automatically. E.g., when there is a column from A1 to A10 with
a colorscale, and the user copies a cell from that range to A11, it
is very likely that the intention is to expand the existing CF, not
to create a separate colorscale individually for the cell A11.

This makes it necessary to create some complexity in deciding when to
deduplicate, and when not. This change introduces a function for that,
called isRangeDependentFormatNeedDeduplication; and implements some
simple algorithm that currently only depends on analysis of the two
CFs' ranges. If both ranges are one-dimensional vectors, and the new
CF range "continues" the existing range (the edges of the two vectors
are adjacent), then it decides to combine them into one expanded CF.
Also, if the new range is completely inside the old one (which may
easily happen, when a user copies a cell from one place to another
inside the same range), it also allows deduplication. In other cases,
it rejects deduplication, requiring creation of a new CF.

It may be extended as needed. One possible improvement could be e.g.
when a user has a two-dimensional colorscale block, and copies some
of its rows below, then it would be likely that the intention is to
expand the range of existing CF, rather than creating a new one for
the expanded range. This would need some specification how to tell
this case from others that must not deduplicate.

I must note that this problem is indeed difficult: e.g., in Excel,
there is also no simple logic around copying such CF. I saw very odd
cases, where the copy resulted in merging ranges of existing CF, or
in creation of new CF; also, it could even magically add cells to
existing CFs without copy - just by entering data below existing CF
(and even when there was some gap between the existing CF range and
the newly entered data - but not when there was two gaps). So there
likely will always be some corner cases where users expect something
else.

This changes the behavior established in commits
3f614f431475e1bf3bb3bbeac59b0681309628b7 (tdf#95295: don't add
duplicate conditional formats, 2017-12-11),
3fa15dd614bd72ddb36dbe033abeef5609d31f38 (tdf#154906 tdf#129813
tdf#129814 sc: fix conditional format color scale, 2023-04-26)
and 8af6c46a9c0e86bbbd908e96ff236ad1d6c4ddab (tdf#155319 sc: fix
conditional format data bar after copying, 2023-05-22) regarding the
mentioned types of CF.

Change-Id: Ie94167bbf66e448a6b56ad0962cb38221aa7d4b7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180735
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins