tdf#145160 Set default choice in Image Compression dialog per last format
Currently the compression format selected by default in the compressed
dialog box for an image depends on the option last selected by the user
to compress an image except when it is opened for the first time after
opening libreoffice in which case the `PNG Compression` option is
selected. This happens regardless of the image format. This patch sets
the default option based on the format of the image each time the dialog
box is opened. If it is a JPG, then `JPEG Quality` will be chosen otherwise
the 'PNG Compression' will be chosen.

Change-Id: Ib8eb84b6b1e226f5bf6e98aa1e0477ffced9be73
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180758
Reviewed-by: Sahil Gautam <sahil.gautam.extern@allotropia.de>
Tested-by: Jenkins