a11y: Drop non-functional bits from AccessibleGridControl::getAccessibleAtPoint
XAccessibleComponent::getAccessibleAtPoint should return
the direct child at the given point.

AccessibleGridControl can have up to 3 children,
see AccessibleGridControl::getAccessibleChild:

* a row header (AccessibleGridControlHeader)
* a column header (AccessibleGridControlHeader)
* the table (AccessibleGridControlTable)

TableControl::ConvertPointToControlIndex on the other
hand returns the child index of a cell within
the table, so could only be relevant when
calling AccessibleGridControlTable::getAccessibleAtPoint on the
AccessibleGridControlTable child
(and AccessibleGridControlTable::getAccessibleAtPoint does similar,
but uses a method called TableControl::ConvertPointToCellAddress
instead).

In addition, TableControl::CreateAccessibleControl that
was called with the index retrieved that way was
warning it should be overriden (Where?) and unconditionally
returning nullptr.

Drop that broken code path. The remaining logic
to iterate over the children looks reasonable in principle.

Without this commit in place, using the "Inspect object under mouse"
feature in Accerciser when the mouse is over a cell of the grid control
of the sample doc attachment 198647 from tdf#164783 with the qt6 VCL plugin
would result in the top-level's TableControl's a11y object getting selected
in Accerciser's treeview of the LO a11y and its full area getting
highlighted.
Now, it selects/highlights the column header cell - even if the mouse cursor
is over one of the regular cells - which is still not fully correct,
but that is a different issue.

Change-Id: Ib4a0d05145c66a96dcb72a6a50805182c17fb2f5
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180792
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins