a11y: Drop IAccessibleTable
Now that all related code is in the toolkit module
and the interface (abstract base class) is no more
needed to break a dependency cycle, drop the interface altogether.

(An earlier version of this change resulted in LO clang plugins
pointing out some oddities. These have in the meantime been addressed
separately in previous commits to prepare for this change.)

Change-Id: I6295695d9f0bf02be26175c45142965351ec5eff
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180797
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins