toolkit a11y: Simplify AccessibleGridControl::getAccessibleAtPoint
In order to check whether any of the children contains
the given point, use the existing
AccessibleGridControl::getAccessibleChild method  to iterate
over the actually existing children instead of having
an additional AccessibleGridControl::implGetFixedChild
that uses static/fixed indices for the different kinds
of potential children that may or may not be there.

Drop now unused helper methods that mostly duplicated
logic already found in AccessibleGridControl::getAccessibleChild.

Change-Id: I57b814aa497097d1f18506303a76d53ec9f6e0b1
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180799
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins