uno grid a11y: Don't confuse row/col
TableControl::calcHeaderRect's `_bIsColumnBar` param is
passed as the `bColHeader` to TableControl_Impl::calcHeaderRect,
both of which have the same meaning ("whether this is a column
header (cell)"), so inverting it (for no apparent reason) is
incorrect.

As a consquence, selecting the grid control's column header bar
in Accerciser's treeview of the LO a11y hierarchy with sample doc
attachment 198647 from tdf#164783 with the qt6 VCL plugin
would result in not only the header area getting highlighted, but
the area of the rest of the table as well (as rectangle returned by
AccessibleGridControlHeaderCell::implGetBoundingBoxOnScreen
result was incorrect).

With this commit in place, the header area is now highlighted
as expected.

Change-Id: I2c99c940faf469678ecee33dd029580e57170f2a
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180802
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>