uno grid a11y: Pass pos relative to TableControl as expected
TableControl::ConvertPointToCellAddress expects a position
relative to the TableControl.
However, AccessibleGridControlTable::getAccessibleAtPoint
gets a position relative to itself, which is different
at least when the table also has row/col headers.

Add the own position relative to the parent (which is the
TableControl) to convert the position as needed.

With this in place, using Accerciser's "Select object under mouse"
feature with the grid control from attachment 198647 from tdf#164783
with the qt6 VCL plugin now results in the correct cell being
highlighted when the mouse pointer is over a regular cell, while
the cell one row above was highlighted previously.

The cell is still not selected in Accerciser's treeview
of LO a11y tree, but that's still another issue.

Change-Id: If9456b6399ad7a871cb52d9ff4f38168e5907579
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180803
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>