tdf#130857 qt weld: Support MenuButton separators + respect pos
Implement QtInstanceMenuButton::insert_separator.
Add a helper method QtInstanceMenuButton::insertAction
that inserts the given action at the given position
and use it from QtInstanceMenuButton::insert_item
as well, which was so far only assigning the `nPos`
local var, but not actually taking it into account
to specify where to insert the new action.

Use QWidget::insertAction [1] that allows to specify before
which existing action the action should be inserted
instead of always appending to the end.

[1] https://doc.qt.io/qt-6/qwidget.html#insertAction

Change-Id: I0acefb29f01b680b4efdf0ad6cea3bbf7f84c4b7
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180807
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>