Introduce EMSCRIPTEN_EXTRA_QTLOADER_CONFIG
for use in Emscripten --enable-qt6 builds, to e.g. add command line arguments
(like --writer) into qt_soffice.html.

For standard --enable-qt5 Emscripten builds, it works to specify an

> EMSCRIPTEN_EXTRA_SOFFICE_PRE_JS=/path/to/include.js

containing

> Module.arguments = ['--writer'];

to make qt_soffice.html open a Writer document.  Due to how the Qt loading has
changed in Qt6, that no longer works for --enable-qt6 Emscripten builds.  But
what works there now is to instead specify an

> EMSCRIPTEN_EXTRA_QTLOADER_CONFIG=arguments:['--writer']

(The sed -z option that is necessary to replace text spanning a newline is a
non-standard GNU extension, but lets assume that this is only executed for
cross-builds from Linux that use GNU sed.)

Change-Id: I31372a7a1f8fccbe0fe2594614bf5846e5e5c95f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180821
Tested-by: Jenkins
Reviewed-by: Stephan Bergmann <stephan.bergmann@allotropia.de>