uno grid a11y: Merge AccessibleGridControl{,Access}
So far, AccessibleGridControlAccess was implementing
the XAccessible interface, and its
XAccessible::getAccessibleContext implementation
was returning a AccessibleGridControl instance
that implements XAccessibleContext.

Drop that separation and let AccessibleGridControl
also implement the XAccessible interface directly
and return itself in
AccessibleGridControl::getAccessibleContext.

TableControl_Impl::getAccessible now creates
an AccessibleGridControl accordingly.

No change in behavior intended or seen when
testing with the grid in the sample doc attachment 198647
from tdf#164783 when testing with Orca and Accerciser
with the gtk3 and qt6 VCL plugins.
Change-Id: Iabc4cd11e7b6b6e121c65a6813a28e218c34b4e1

Change-Id: I8bfdf64f436a831251f155d0e1238e4d82d1fa58
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180832
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>