uno grid a11y: Unify XAccessible implementation in base class
By now, all AccessibleGridControlBase subclasses implement
both, the XAccessible and XAccessibleContext interfaces,
and their XAccessible::getAccessibleContext implementations
return a reference to self.

This is implemented in different ways: Either implementing
it directly or using the GridControlAccessibleElement
class that does it as base class.

Unify and simplify this by letting AccessibleGridControlBase
implement the XAccessible interface, and dropping the extra
handling in all subclasses.

(Other than the other implementations,
AccessibleGridControlHeaderCell::getAccessibleContext
was previously not using SolarMutexGuard, but I can't
see a particular reason, so don't see a reason to
override the new base class implementation.)

Change-Id: Ibdb674c5d5556fcf1c5644a2f6d961985b0e3362
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180838
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>
Tested-by: Jenkins