a11y: Use OAccessibleComponentHelper for AccessibleListBoxEntry
Similar to how

    commit 7107db3fe773629cc511eb5922bc9c28c9c5c60a
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jan 24 18:57:58 2025 +0100

        toolkit a11y: Use OAccessibleComponentHelper for grid control

did for AccessibleGridControlBase, also derive
from OAccessibleComponentHelper for AccessibleListBoxEntry,
to make use of the logic already implemented there
instead of having to implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods  manually

AccessibleListBoxEntry::GetBoundingBox_Impl already
provides what is needed to implement
AccessibleListBoxEntry::implGetBounds, just needs
to convert to awt::Rectangle.

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the OAccessibleComponentHelper base class.

No change in behavior intended or observed when
testing this with Accerciser and Orca and the qt6
VCL plugin with the "Tools" -> "AutoText" dialog
in Writer, where this code path is used for
items in the treeview.

Change-Id: I5c95bc6135db92a6e9d736a134af2cbcc35a5e06
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180840
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>