a11y: Use OAccessibleComponentHelper for text window Paragraph
Similar to how

    commit 7107db3fe773629cc511eb5922bc9c28c9c5c60a
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Fri Jan 24 18:57:58 2025 +0100

        toolkit a11y: Use OAccessibleComponentHelper for grid control

did for AccessibleGridControlBase, also derive
from OAccessibleComponentHelper for Paragraph
to make use of the logic already implemented there
instead of having to implement all the XAccessibleComponent
and XAccessibleEventBroadcaster methods manually.

Paragraph::getBounds already
provides what is needed to implement
Paragraph::implGetBounds, so just rename it.

Drop all of the other overrides that are no longer
needed as the implementation is now provided by
the OAccessibleComponentHelper base class.

No change in behavior intended or observed when
testing this with Accerciser and Orca and the qt6
VCL plugin with the IDE editor ("Tools" -> "Macros"
-> "Edit Macros").

Change-Id: I88ac869d115ceb13ccf48932f6d66d95bbd2fc9b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180896
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>