sw: layout: fix loop with table at the end of a section
There is a table alone in a section, and there are 2 rows in the table,
with the 2nd row split across page 1 and page 2.

Due to space for another line of text being available, the table frame
first does RemoveFollowFlowLine() etc and grows much higher than the
section frame upper.

Then in SwTabFrame::Join(), the section frame on page 2 is destroyed via
SwSectionFrame::DelEmpty() so the section frame on page 1 doesn't have a
follow.

Then a SwTabFrame::Split() happens, and lcl_ShrinkCellsAndAllContent()
ends up shrinking the section frame to 0 height (because there is no
follow currently seen by SwSectionFrame::ToMaximize()), which is wrong
as there is a 1st row in the table, so the table is still height 3712.

If the table is directly in the body, then this works because the body
frame doesn't shrink or grow.

Then when the text frame in the 2nd row is formatted, it can't grow
because SetRestrictTableGrowth(true) was called as a follow flow line
exists, and in SwTabFrame::GrowFrame() nReal is -2777 due to the
SwTabFrame being higher than the upper; the result is that the text
frame is the same height/split at the same line as before
SwTabFrame::MakeAll().

Then the whole thing happens again until eventually
warn:sw.layout:875613:875613:sw/source/core/layout/layouter.cxx:194: Looping Louie: Stage 1!

Try to detect situation of too high table in section and propagate
Shrink only until the section is the same height as the table.

Change-Id: If4b6db8b03d93a3227ad223203f2b0c3f73d3cc8
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180857
Tested-by: Jenkins
Reviewed-by: Michael Stahl <michael.stahl@allotropia.de>