a11y: Use existing method to get VCLXAccessibleComponent parent
Call `getAccessibleParent()` to
get the parent instead of reimplementing the logic
from VCLXAccessibleComponent::getAccessibleParent
in VCLXAccessibleComponent::getAccessibleIndexInParent.

Calling `getAccessibleParent()` could give a different
result when any subclass overrides
VCLXAccessibleComponent::getAccessibleParent. But the
logic implemented here should use the actual accessible
parent, so if it didn't use the actual parent for any
subclass before this commit, that was probably a bug that
this commit should fix as well.

(`git show --ignore-space-change` makes the "actual
change" clearer.)

Change-Id: I1d7c43d5ffa7abb8546ad50b7c6a4f9773d5c583
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180868
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>