sw: fix bad layout due to stale SwLayouter
The previous commit still leaves a problem on this branch, which happens
after toggling Ctrl+F9 a dozen or so times, the layout is broken with:

warn:legacy.osl:1814070:1814070:sw/source/core/layout/flowfrm.cxx:2531: <SwFlowFrame::MoveBwd(..)> - layout loop control for layout action <Move Backward> applied!

The problem is that SwLayouter contains stale entries from the previous
layout action that need to be cleared before starting the next one.

This is part of master commit b0e2b60fa45f236f6a993cc1295a7afddabb5098
but that is too risky to backport as a whole.

Change-Id: Ib4ac6345acd1e94370dd799e685a4c378228d908