tdf#164885: use SfxBindings::Execute instead of SfxDispatcher::ExecuteList
Commit 502f7958d99e3aa27f4c1c1f33f4827055459da8 (pass a frame to use as
parent for any dialog that may need to appear, 2023-11-16) switched to
SfxDispatcher::ExecuteList, in order to pass there frame to be used as
parent for dialogs.

Later, in commit d7b33b1be8fee684e906f170f4001735c095056e (use SID_DOCFRAME
for dialog parents if there is no SID_FILLFRAME, 2023-11-23), the fix
was changed, and passing the frame got not needed.

Since the change to SfxDispatcher::ExecuteList caused the regression
(likely because SfxDispatcher unexpectedly uses a bit different method
to find the slot server), and the original reason for the change is no
more actual, let's revert to the older method.

Change-Id: I6b126bf336dcf8cd3f49a3ce9682f3adc3bc208f
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180909
Reviewed-by: Mike Kaganski <mike.kaganski@collabora.com>
Tested-by: Jenkins