icon choice ctrl a11y: Don't announce selection change on focus
When processing a VclEventId::WindowGetFocus event,
only send and event to notify that the active
descendant (= child that has focus) of the icon choice control
has changed, but don't announce a selection change event.
The latter is handled in VclEventId::ListboxSelect instead.
(And when receiving focus, the selected entry generally remains
the same, i.e. the selection does not change.)

Also, don't try to extract the selected entry
from the data sent in the focus event.

This code gets triggered for the "Insert" -> "Hyperlink"
dialog in Writer.

Change-Id: I013da234b78f872b1b286769acdf38a66d58801b
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180911
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>