icon choice ctrl a11y: Separate selection from active descendant change
When receiving a VclEventId::ListboxSelect event,
always send an a11y event to notify about the changed
selection if an entry is selected, regardless of whether
or not the icon choice control has focus.

Make only the ACTIVE_DESCENDANT_CHANGED event
conditional on focus, as that is the event that
notifies about the child that newly receives focus.

Change-Id: Ibea29092ff33be69fde098777b13a9fa96b19c8e
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180912
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>