icon choice ctrl a11y: Don't announce selection change on focus (II)
    Change-Id: I013da234b78f872b1b286769acdf38a66d58801b
    Author: Michael Weghorn <m.weghorn@posteo.de>
    Date:   Wed Jan 29 16:45:23 2025 +0100

        icon choice ctrl a11y: Don't announce selection change on focus

had implemented not "inventing" a selection change event
in the a11y implementation, AccessibleIconChoiceCtrl,
when the icon choice control receives focus.

However, the SvtIconChoiceCtrl itself was also sending
a VclEventId::ListboxSelect event that gets processed
by the AccessibleIconChoiceCtrl, which again resulted
in an a11y selection changed event getting emitted
(`object:selection-changed` for AT-SPI2).

No longer send that event to not notify of a fake
selection change. (From all I can see,
AccessibleIconChoiceCtrl is the only consumer
of that event.)

Actual changes in the selected entry are still
announced fine (see SvxIconChoiceCtrl_Impl::EntrySelected).

With the qt6 VCL plugin on Linux and the Orca screen
reader, this makes Orca now announce "List with 4 items"
before it announces the currently selected item
when focus moves to the icon choice control
("vertical tabbar") in the "Insert" -> "Hyperlink"
dialog in Writer. (Previously, only the accessible
name of the selected item was announced, but not
the list it is contained in.)

Change-Id: I0a49a034f37fb4fb3a7ea6953c2e545fec184f84
Reviewed-on: https://gerrit.libreoffice.org/c/core/+/180913
Tested-by: Jenkins
Reviewed-by: Michael Weghorn <m.weghorn@posteo.de>